requirements.txt
anki==2.1.26
ankirspy==2.1.26
appdirs==1.4.4
aqt==2.1.26
arandr==0.1.10
asyncio==3.4.3
attrs==20.1.0
beautifulsoup4==4.9.1
browsermob-proxy==0.8.0
bs4==0.0.1
btrfsutil==1.2.0
CacheControl==0.12.6
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0.post1
cycler==0.10.0
Cython==0.29.14
decorator==4.4.2
dill==0.3.1.1
distlib==0.3.1
distro==1.5.0
docopt==0.6.2
filelock==3.0.12
flawfinder==2.0.11
gitdb2==2.0.6
GitPython==3.0.5
html5lib==1.1
idna==2.10
importlib-metadata==1.6.1
joblib==0.14.1
jsonschema==3.2.0
kiwisolver==1.1.0
lensfun==0.3.95
lockfile==0.12.2
Markdown==3.2.2
matplotlib==3.1.3
more-itertools==8.5.0
msgpack==1.0.0
numpy==1.17.4
ordered-set==4.0.2
packaging==20.4
pandas==0.25.3
pep517==0.8.2
pew==1.2.0
pipenv==2020.8.13
praw==7.0.0
prawcore==1.4.0
progress==1.5
protobuf==3.12.4
psutil==5.7.2
pulsemixer==1.5.1
PyAudio==0.2.11
pycairo==1.19.1
pycvesearch==0.5
pyee==7.0.2
PyGObject==3.36.1
pyparsing==2.4.7
pyppeteer==0.2.2
pyppeteer-stealth==1.0.0
PyQt5==5.15.0
PyQt5-sip==12.8.1
PyQtWebEngine==5.15.0
pyreadr==0.2.1
pyrsistent==0.16.0
python-dateutil==2.8.1
python-utils==2.3.0
python-xlib==0.27
pytoml==0.1.21
pytz==2019.3
pywal==3.3.0
PyYAML==5.3.1
requests==2.24.0
resolvelib==0.4.0
retrying==1.3.3
scikit-learn==0.22.2.post1
scipy==1.4.1
selenium==3.141.0
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
smmap2==2.0.5
soupsieve==2.0.1
style==1.1.0
team==1.0
TikTokApi==3.4.6
toml==0.10.1
tqdm==4.46.1
ueberzug==18.1.6
update==0.0.1
update-checker==0.17
urllib3==1.25.10
urlscan==0.9.5
urwid==2.1.1
utils==0.9.0
vccrosshair @ file:///home/rappy/Programs/Repos/vccrosshair
virtualenv==20.0.23
virtualenv-clone==0.5.4
webencodings==0.5.1
websocket-client==0.57.0
websockets==8.1
youtube-dl==2020.7.28
zipp==3.1.0
